Correct project structure
Challenger is a binary crate, and is not intended to be used as a
library. As such, lib.rs has been deleted, and it's content moved to a
more appropriate place within the uci.rs module.
